+++ tests/test_vfpproperties.cpp
@@ -39,7 +39,9 @@
 #include <opm/parser/eclipse/EclipseState/checkDeck.hpp>
 #include <opm/parser/eclipse/EclipseState/EclipseState.hpp>
 #include <opm/parser/eclipse/EclipseState/Tables/VFPProdTable.hpp>
+
 #include <opm/autodiff/VFPProperties.hpp>
+#include <opm/autodiff/VFPProdProperties.hpp>
